requirements.txt
# Core Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0

# Git Operations
GitPython>=3.1.40
pygit2>=1.13.3

# GitHub API
PyGithub>=2.1.1
requests>=2.31.0

# Database
SQLAlchemy>=2.0.23

# Code Analysis
radon>=6.0.1
lizard>=1.17.10
pydriller>=2.5

# LLM Integration
openai>=1.6.1
anthropic>=0.8.1

# Data Processing
pandas>=2.1.4
numpy>=1.26.2

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
colorama>=0.4.6

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.12.1
flake8>=6.1.0
mypy>=1.7.1